% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dothis.R
\docType{package}
\name{dothis}
\alias{dothis}
\alias{dothis-package}
\title{dothis: a very basic "To Do" List for RStudio}
\description{
This package adds 2 items to the Addins menu in RStudio and provides a very
basic "To Do" List, inspired by the task pane in Eclipse.
}
\details{
All comments starting with the keywords "todo" or "fixme" will be listed in
the Markers pane.
}

