% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_todolist.R
\name{generate_todolistAddin}
\alias{generate_todolistAddin}
\title{Generate a "To Do" List in the markers pane}
\usage{
generate_todolistAddin()
}
\value{
returns invisibly
}
\description{
This function should never be called directly.
}

