% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_todolist.R
\name{generate_todolistAddin}
\alias{generate_todolistAddin}
\title{Generate a "To Do" List in the Rstudio markers pane}
\usage{
generate_todolistAddin()
}
\value{
returns invisibly
}
\description{
This function looks for comments in the active document starting with the
words 'todo' or 'fixme' and gathers them in the markers pane. This function
should never be called directly - it is added as an item in the Addins menu.
}

